Factor out falloff functions from proportional.py to math.py.
