+++ utils/math.py
@@ -1,0 +1,49 @@
+# ##### BEGIN GPL LICENSE BLOCK #####
+#
+#  This program is free software; you can redistribute it and/or
+#  modify it under the terms of the GNU General Public License
+#  as published by the Free Software Foundation; either version 2
+#  of the License, or (at your option) any later version.
+#
+#  This program is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU General Public License for more details.
+#
+#  You should have received a copy of the GNU General Public License
+#  along with this program; if not, write to the Free Software Foundation,
+#  Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
+#
+# ##### END GPL LICENSE BLOCK #####
+
+import math
+
+def smooth(x):
+    return 3*x*x - 2*x*x*x
+
+def sharp(x):
+    return x * (2 - x)
+
+def root(x):
+    return 1.0 - math.sqrt(1.0 - x)
+
+def linear(x):
+    return x
+
+def const(x):
+    return 0.0
+
+def sphere(x):
+    return 1.0 - math.sqrt(1.0 - x*x)
+
+def invsquare(x):
+    return x*x
+
+def falloff(type, radius, rho):
+    if rho <= 0:
+        return 1.0
+    if rho > radius:
+        return 0.0
+    func = globals()[type]
+    return 1.0 - func(rho / radius)
+
